prepare async addBook
