+++ src/redux/books/booksSlice.js
@@ -1,54 +1,9 @@
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
+import { createSlice } from '@reduxjs/toolkit';
 import unique from '../../util/unique';
-
-const URL = 'https://us-central1-bookstore-api-e63c8.cloudfunctions.net/bookstoreApi/apps/HYYvKmGO3upxaXekhp4l/books';
-// const delURL = (id) => `${URL}/id`;
-
-export const fetchBooks = createAsyncThunk('books/fetchBooks', async () => {
-  // console.log('fetchbooks is called');
-  const response = await fetch(URL)
-    .then((res) => {
-      if (!res.ok) {
-        throw new Error(`HTTP error! Status: ${res.status}`);
-      }
-      return res.json();
-    })
-    .catch((err) => err.message);
-
-  // const list = unique(response, );
-
-  return response;
-});
-
-// fetch(addUrl, {
-//   method: 'POST',
-//   headers: {
-//   "Content-Type": "application/json",
-// },
-//   body: JSON.stringify(Book)
-// })
+import { fetchBooks } from './booksThunks';
 
 const initialState = {
-  books: [
-    {
-      item_id: 'item1',
-      title: 'The Great Gatsby',
-      author: 'John Smith',
-      category: 'Fiction',
-    },
-    {
-      item_id: 'item2',
-      title: 'Anna Karenina',
-      author: 'Leo Tolstoy',
-      category: 'Fiction',
-    },
-    {
-      item_id: 'item3',
-      title: 'The Selfish Gene',
-      author: 'Richard Dawkins',
-      category: 'Nonfiction',
-    },
-  ],
+  books: [],
   status: 'idle', // 'idle', 'loading', 'succeeded', 'failed'
   error: null,
 };
@@ -72,8 +27,7 @@ const booksSlice = createSlice({
       .addCase(fetchBooks.pending, (state) => ({ ...state, status: 'loading' }))
       .addCase(fetchBooks.fulfilled, (state, action) => ({
         ...state,
-        // TODO: handle extra redundant render
-        // console.log(action.payload)
+        // TODO: handle (should be redundant) render
         books: unique(action.payload, state.books),
         status: 'succeeded',
       }))
@@ -84,7 +38,7 @@ const booksSlice = createSlice({
       }));
   },
 });
-
+export { fetchBooks };
 export const booksState = (state) => state.books;
 export const { removeBook, addBook } = booksSlice.actions;
 export default booksSlice.reducer;
