+++ src/util/unique.js
@@ -4,6 +4,5 @@ export default (res, current) => {
     .filter(([key]) => !current.map((item) => item.item_id).includes(key))
     // reshaping the array and objects according to design scheme
     .map(([key, [value]]) => ({ item_id: key, ...value }))];
-  console.log(arr.concat(current));
   return arr;
 };
